feat: update history event utils, separate history event composables
